Renaming issue fixed to reduce diff noise
Renaming fixed
